refactor into pkg
